fix(S7b): MiMo review follow-up on edb7ffa — parity-test overflow + odd_ext guard
Verified findings from the MiMo review of the normalization core:

1. Parity-test delta arithmetic widened to i64 (BUG). The golden Q31 values
   span ~±1.5e9, so `(r - g)` for a large divergence could exceed i32 range
   (in debug it panics; in release it wraps) and mask a real failure. `max_delta`
   and `TOL` are now i64 and the diff computes as `r as i64 - g as i64`. Gate
   still reports max|Δ|=0 (bit-exact).

2. `odd_ext` precondition made explicit (STYLE). Added a `debug_assert!(x.len()
   > n)`. NB the reviewer flagged it as a `pub` OOB risk — that's a false
   positive (odd_ext is a private `fn`, and its sole caller `sosfiltfilt_hp`
   already guards the length); the assert just documents the contract for a
   future caller.

3. Noted that the hardcoded SOS coeffs are guarded by the
   `sosfiltfilt_matches_scipy_oracle_on_ramp20` test (wrong coeffs → the oracle
   comparison diverges), since there is no runtime Rust `butter` to re-derive
   them from HIGHPASS_HZ/TARGET_SR.

Skipped MiMo findings — deferred / non-issues (audit trail):
- q31_normalize `&[Vec<f64>]` API shape: deferred to the resample increment
  (the reviewer's own recommendation) where the buffer layout may change.
- None-on-flat contract: already documented (= Python's skip); the eventual
  pyfunction wrapper will match the Python `except` skip behavior.
- TOL=2 unused headroom / scipy drift: the golden is committed + frozen (only
  the dump script regenerates it) and the manifest records the scipy version,
  so a scipy upgrade cannot silently widen the gap.
- Critic "trim assumes rev.len() >= 2n" and "f32 vs f64 input": the reviewer
  withdrew both on closer inspection (guaranteed by construction; lossless upcast).

3 normalize unit tests + parity gate green (still bit-exact); core-no_std
untouched (host-only module).

+++ lamquant-lossless/tests/normalize_parity.rs
@@ -25,7 +25,8 @@ const T: usize = 128;
 /// Documented tolerance: max |Δ| between the Rust and scipy int32 Q31 outputs.
 /// Both run the identical f64 filtfilt algorithm, so this is FP-order slack at
 /// the truncation boundary, not an algorithmic gap. Tighten if it ever drifts.
-const TOL: i32 = 2;
+/// i64 so the delta of two near-±i32::MAX Q31 values can't overflow (MiMo review).
+const TOL: i64 = 2;
 
 /// MUST match `make_input` in `tools/dump_normalize_golden.py`.
 fn synth_input() -> Vec<Vec<f64>> {
@@ -69,11 +70,11 @@ fn rust_normalization_matches_python_golden_within_tolerance() {
     let gold = golden();
     assert_eq!(flat.len(), gold.len());
 
-    let mut max_delta = 0i32;
+    let mut max_delta = 0i64;
     let mut n_mismatch = 0usize;
     let mut worst_at = 0usize;
     for (i, (&r, &g)) in flat.iter().zip(gold.iter()).enumerate() {
-        let d = (r - g).abs();
+        let d = (r as i64 - g as i64).abs();
         if d > 0 {
             n_mismatch += 1;
         }
